fix(agent): share the runaway guard across every front end, log ACP sessions
The stuck detector was instantiated only in internal/tui/agent_loop.go, so
only the TUI was guarded. runPrint, json, rpc and ACP ran with no runaway
protection at all: replay trials of a known degenerate turn produced 71, 67
and 42 byte-exact repeats and every one ran to completion without aborting,
against a threshold of 12.

Move StuckDetector, its thresholds and helpers to internal/agent/stuck.go and
give it one entry point, ObserveEvent, which feeds an event's parts to all
three arms in a fixed order. The TUI, print and json modes and the ACP prompt
turn now all call it, so no front end can drift from the others or silently
skip the guard. Behaviour and thresholds are unchanged.

ACP additionally wrote no session log at all, which left its runs invisible to
everything that reads ~/.pi-go/log -- so a loop there was both unguarded and
undiagnosable after the fact. It now creates a session log like every other
mode, records session_start with mode "acp-server", and logs thinking, text,
tool calls and tool results in the same entry shape. A log that cannot be
created is not fatal: the server still serves, and logEventParts is a no-op on
a nil logger.

Tests: five for ObserveEvent covering each arm, the nil/empty no-op and that
productive polling is not mistaken for a loop; two for the ACP log asserting
every part type is written, that tool results are JSON rather than a %v dump,
and that a nil logger writes nothing. The detector's unit tests move to
internal/agent with it; the TUI keeps its integration tests and refers to the
thresholds through the shared package.
